gateway-service/src/test/java/com/adidas/challenge/gateway/auth/TokenFilterTest.java
package com.adidas.challenge.gateway.auth;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.test.context.junit4.SpringRunner;

import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import java.io.IOException;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Mockito.atLeast;
import static org.mockito.Mockito.verify;

@RunWith(SpringRunner.class)
@SpringBootTest(properties = { "jwt.header=Authorization" })
public class TokenFilterTest {

    @InjectMocks
    private TokenFilter tokenFilter;

    @Mock
    private FilterChain filterChain;

    private MockHttpServletRequest request;
    private MockHttpServletResponse response;

    @Before
    public void setUp() {
        request = new MockHttpServletRequest();
        response = new MockHttpServletResponse();
    }

    @Test
    public void shouldAuthenticate() throws IOException, ServletException {
        // given
        request.setServletPath("/subscription");
        request.addHeader("Authorization", 1234567890);

        // when
        tokenFilter.doFilter(request, response, filterChain);

        // then
        UsernamePasswordAuthenticationToken authenticationToken = (UsernamePasswordAuthenticationToken) SecurityContextHolder.getContext().getAuthentication();
        verify(filterChain, atLeast(1)).doFilter(request, response);
        assertThat(authenticationToken).isNotNull();
    }
}
